TURMA 37 GENERATION

+++ blogPessoal (3)/blogPessoal/src/main/java/org/generation/blogPessoal/controller/PostagemController.java
@@ -2,6 +2,8 @@ package org.generation.blogPessoal.controller;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.generation.blogPessoal.model.Postagem;
 import org.generation.blogPessoal.repository.PostagemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +17,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/postagens")
@@ -44,17 +48,26 @@ public class PostagemController {
 	public ResponseEntity<List<Postagem>> GetByTexto(@PathVariable String texto) {
 		return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(texto));
 	}
-	@PostMapping 
-	public ResponseEntity<Postagem> post(@RequestBody Postagem postagem){
-		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
+	@PostMapping
+	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem){
+	        return ResponseEntity.status(HttpStatus.CREATED)
+			        .body(repository.save(postagem));
 	}
-	@PutMapping 
-	public ResponseEntity<Postagem> put(@RequestBody Postagem postagem){
-		return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+
+	@PutMapping
+	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem){
+	        return repository.findById(postagem.getId())
+			        .map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+			        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
+
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable Long id) {
-		repository.deleteById(id);
-	}
+	public void delete(@PathVariable long id) {
+	        java.util.Optional<Postagem> post = repository.findById(id);
+	        if(post.isEmpty())
+		        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+	        repository.deleteById(id);				
 	}
 	
+}
